+++ HPXMLtoOpenStudio/resources/hpxml_defaults.rb
@@ -1295,6 +1295,17 @@ class HPXMLDefaults
       end
     end
 
+    # Pilot Light
+    hpxml.heating_systems.each do |heating_system|
+      if heating_system.pilot_light.nil?
+        heating_system.pilot_light = false
+        heating_system.pilot_light_isdefaulted = true
+      elsif heating_system.pilot_light && heating_system.pilot_light_btuh.nil?
+        heating_system.pilot_light_btuh = 500.0
+        heating_system.pilot_light_btuh_isdefaulted = true
+      end
+    end
+
     # Detailed HVAC performance
     hpxml.cooling_systems.each do |cooling_system|
       clg_ap = cooling_system.additional_properties