+++ HPXMLtoOpenStudio/tests/test_defaults.rb
@@ -1158,17 +1158,26 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml.heating_systems[0].fan_watts_per_cfm = 0.66
     hpxml.heating_systems[0].airflow_defect_ratio = -0.22
     hpxml.heating_systems[0].heating_capacity = 12345
+    hpxml.heating_systems[0].pilot_light = true
+    hpxml.heating_systems[0].pilot_light_btuh = 999
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_furnace_values(hpxml_default.heating_systems[0], 0.66, -0.22, 12345)
+    _test_default_furnace_values(hpxml_default.heating_systems[0], 0.66, -0.22, 12345, true, 999)
 
     # Test defaults
     hpxml.heating_systems[0].fan_watts_per_cfm = nil
     hpxml.heating_systems[0].airflow_defect_ratio = nil
     hpxml.heating_systems[0].heating_capacity = nil
+    hpxml.heating_systems[0].pilot_light_btuh = nil
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_furnace_values(hpxml_default.heating_systems[0], 0.375, 0, nil)
+    _test_default_furnace_values(hpxml_default.heating_systems[0], 0.375, 0, nil, true, 500)
+
+    # Test defaults w/o pilot
+    hpxml.heating_systems[0].pilot_light = nil
+    XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
+    hpxml_default = _test_measure()
+    _test_default_furnace_values(hpxml_default.heating_systems[0], 0.375, 0, nil, false, nil)
 
     # Test defaults w/ gravity distribution system
     hpxml = _create_hpxml('base-hvac-furnace-gas-only.xml')
@@ -1178,7 +1187,7 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml.heating_systems[0].heating_capacity = nil
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_furnace_values(hpxml_default.heating_systems[0], 0.0, 0, nil)
+    _test_default_furnace_values(hpxml_default.heating_systems[0], 0.0, 0, nil, false, nil)
   end
 
   def test_wall_furnaces
@@ -1186,16 +1195,25 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml = _create_hpxml('base-hvac-wall-furnace-elec-only.xml')
     hpxml.heating_systems[0].fan_watts = 22
     hpxml.heating_systems[0].heating_capacity = 12345
+    hpxml.heating_systems[0].pilot_light = true
+    hpxml.heating_systems[0].pilot_light_btuh = 999
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_wall_furnace_values(hpxml_default.heating_systems[0], 22, 12345)
+    _test_default_wall_furnace_values(hpxml_default.heating_systems[0], 22, 12345, true, 999)
 
     # Test defaults
     hpxml.heating_systems[0].fan_watts = nil
     hpxml.heating_systems[0].heating_capacity = nil
+    hpxml.heating_systems[0].pilot_light_btuh = nil
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_wall_furnace_values(hpxml_default.heating_systems[0], 0, nil)
+    _test_default_wall_furnace_values(hpxml_default.heating_systems[0], 0, nil, true, 500)
+
+    # Test defaults w/o pilot
+    hpxml.heating_systems[0].pilot_light = nil
+    XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
+    hpxml_default = _test_measure()
+    _test_default_wall_furnace_values(hpxml_default.heating_systems[0], 0, nil, false, nil)
   end
 
   def test_floor_furnaces
@@ -1203,15 +1221,24 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml = _create_hpxml('base-hvac-floor-furnace-propane-only.xml')
     hpxml.heating_systems[0].fan_watts = 22
     hpxml.heating_systems[0].heating_capacity = 12345
+    hpxml.heating_systems[0].pilot_light = true
+    hpxml.heating_systems[0].pilot_light_btuh = 999
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_floor_furnace_values(hpxml_default.heating_systems[0], 22, 12345)
+    _test_default_floor_furnace_values(hpxml_default.heating_systems[0], 22, 12345, true, 999)
 
     # Test defaults
     hpxml.heating_systems[0].fan_watts = nil
+    hpxml.heating_systems[0].pilot_light_btuh = nil
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_floor_furnace_values(hpxml_default.heating_systems[0], 0, nil)
+    _test_default_floor_furnace_values(hpxml_default.heating_systems[0], 0, nil, true, 500)
+
+    # Test defaults w/o pilot
+    hpxml.heating_systems[0].pilot_light = nil
+    XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
+    hpxml_default = _test_measure()
+    _test_default_floor_furnace_values(hpxml_default.heating_systems[0], 0, nil, false, nil)
   end
 
   def test_boilers
@@ -1219,30 +1246,42 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml = _create_hpxml('base-hvac-boiler-gas-only.xml')
     hpxml.heating_systems[0].electric_auxiliary_energy = 99.9
     hpxml.heating_systems[0].heating_capacity = 12345
+    hpxml.heating_systems[0].pilot_light = true
+    hpxml.heating_systems[0].pilot_light_btuh = 999
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_boiler_values(hpxml_default.heating_systems[0], 99.9, 12345)
+    _test_default_boiler_values(hpxml_default.heating_systems[0], 99.9, 12345, true, 999)
 
     # Test defaults w/ in-unit boiler
     hpxml.heating_systems[0].electric_auxiliary_energy = nil
     hpxml.heating_systems[0].heating_capacity = nil
+    hpxml.heating_systems[0].pilot_light_btuh = nil
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_boiler_values(hpxml_default.heating_systems[0], 170.0, nil)
+    _test_default_boiler_values(hpxml_default.heating_systems[0], 170.0, nil, true, 500)
+
+    # Test defaults w/ in-unit boiler and w/o pilot
+    hpxml.heating_systems[0].pilot_light = nil
+    XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
+    hpxml_default = _test_measure()
+    _test_default_boiler_values(hpxml_default.heating_systems[0], 170.0, nil, false, nil)
 
     # Test inputs not overridden by defaults (shared boiler)
     hpxml = _create_hpxml('base-bldgtype-multifamily-shared-boiler-only-baseboard.xml')
     hpxml.heating_systems[0].shared_loop_watts = nil
     hpxml.heating_systems[0].electric_auxiliary_energy = 99.9
+    hpxml.heating_systems[0].pilot_light = true
+    hpxml.heating_systems[0].pilot_light_btuh = 999
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_boiler_values(hpxml_default.heating_systems[0], 99.9, nil)
+    _test_default_boiler_values(hpxml_default.heating_systems[0], 99.9, nil, true, 999)
 
     # Test defaults w/ shared boiler
     hpxml.heating_systems[0].electric_auxiliary_energy = nil
+    hpxml.heating_systems[0].pilot_light_btuh = nil
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_boiler_values(hpxml_default.heating_systems[0], 220.0, nil)
+    _test_default_boiler_values(hpxml_default.heating_systems[0], 220.0, nil, true, 500)
   end
 
   def test_stoves
@@ -1250,16 +1289,25 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml = _create_hpxml('base-hvac-stove-oil-only.xml')
     hpxml.heating_systems[0].fan_watts = 22
     hpxml.heating_systems[0].heating_capacity = 12345
+    hpxml.heating_systems[0].pilot_light = true
+    hpxml.heating_systems[0].pilot_light_btuh = 999
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_stove_values(hpxml_default.heating_systems[0], 22, 12345)
+    _test_default_stove_values(hpxml_default.heating_systems[0], 22, 12345, true, 999)
 
     # Test defaults
     hpxml.heating_systems[0].fan_watts = nil
     hpxml.heating_systems[0].heating_capacity = nil
+    hpxml.heating_systems[0].pilot_light_btuh = nil
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_stove_values(hpxml_default.heating_systems[0], 40, nil)
+    _test_default_stove_values(hpxml_default.heating_systems[0], 40, nil, true, 500)
+
+    # Test defaults w/o pilot
+    hpxml.heating_systems[0].pilot_light = nil
+    XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
+    hpxml_default = _test_measure()
+    _test_default_stove_values(hpxml_default.heating_systems[0], 40, nil, false, nil)
   end
 
   def test_portable_heaters
@@ -1301,16 +1349,25 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml = _create_hpxml('base-hvac-fireplace-wood-only.xml')
     hpxml.heating_systems[0].fan_watts = 22
     hpxml.heating_systems[0].heating_capacity = 12345
+    hpxml.heating_systems[0].pilot_light = true
+    hpxml.heating_systems[0].pilot_light_btuh = 999
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_fireplace_values(hpxml_default.heating_systems[0], 22, 12345)
+    _test_default_fireplace_values(hpxml_default.heating_systems[0], 22, 12345, true, 999)
 
     # Test defaults
     hpxml.heating_systems[0].fan_watts = nil
     hpxml.heating_systems[0].heating_capacity = nil
+    hpxml.heating_systems[0].pilot_light_btuh = nil
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_fireplace_values(hpxml_default.heating_systems[0], 0, nil)
+    _test_default_fireplace_values(hpxml_default.heating_systems[0], 0, nil, true, 500)
+
+    # Test defaults w/o pilot
+    hpxml.heating_systems[0].pilot_light = nil
+    XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
+    hpxml_default = _test_measure()
+    _test_default_fireplace_values(hpxml_default.heating_systems[0], 0, nil, false, nil)
   end
 
   def test_air_source_heat_pumps
@@ -3616,7 +3673,8 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     end
   end
 
-  def _test_default_furnace_values(heating_system, fan_watts_per_cfm, airflow_defect_ratio, heating_capacity)
+  def _test_default_furnace_values(heating_system, fan_watts_per_cfm, airflow_defect_ratio, heating_capacity,
+                                   pilot_light, pilot_light_btuh)
     assert_equal(fan_watts_per_cfm, heating_system.fan_watts_per_cfm)
     assert_equal(airflow_defect_ratio, heating_system.airflow_defect_ratio)
     if heating_capacity.nil?
@@ -3624,41 +3682,71 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     else
       assert_equal(heating_system.heating_capacity, heating_capacity)
     end
+    assert_equal(heating_system.pilot_light, pilot_light)
+    if pilot_light_btuh.nil?
+      assert_nil(heating_system.pilot_light_btuh)
+    else
+      assert_equal(heating_system.pilot_light_btuh, pilot_light_btuh)
+    end
   end
 
-  def _test_default_wall_furnace_values(heating_system, fan_watts, heating_capacity)
+  def _test_default_wall_furnace_values(heating_system, fan_watts, heating_capacity, pilot_light, pilot_light_btuh)
     assert_equal(fan_watts, heating_system.fan_watts)
     if heating_capacity.nil?
       assert(heating_system.heating_capacity > 0)
     else
       assert_equal(heating_system.heating_capacity, heating_capacity)
     end
+    assert_equal(heating_system.pilot_light, pilot_light)
+    if pilot_light_btuh.nil?
+      assert_nil(heating_system.pilot_light_btuh)
+    else
+      assert_equal(heating_system.pilot_light_btuh, pilot_light_btuh)
+    end
   end
 
-  def _test_default_floor_furnace_values(heating_system, fan_watts, heating_capacity)
+  def _test_default_floor_furnace_values(heating_system, fan_watts, heating_capacity, pilot_light, pilot_light_btuh)
     assert_equal(fan_watts, heating_system.fan_watts)
     if heating_capacity.nil?
       assert(heating_system.heating_capacity > 0)
     else
       assert_equal(heating_system.heating_capacity, heating_capacity)
     end
+    assert_equal(heating_system.pilot_light, pilot_light)
+    if pilot_light_btuh.nil?
+      assert_nil(heating_system.pilot_light_btuh)
+    else
+      assert_equal(heating_system.pilot_light_btuh, pilot_light_btuh)
+    end
   end
 
-  def _test_default_boiler_values(heating_system, eae, heating_capacity)
+  def _test_default_boiler_values(heating_system, eae, heating_capacity, pilot_light, pilot_light_btuh)
     assert_equal(eae, heating_system.electric_auxiliary_energy)
     if heating_capacity.nil?
       assert(heating_system.heating_capacity > 0)
     else
       assert_equal(heating_system.heating_capacity, heating_capacity)
     end
+    assert_equal(heating_system.pilot_light, pilot_light)
+    if pilot_light_btuh.nil?
+      assert_nil(heating_system.pilot_light_btuh)
+    else
+      assert_equal(heating_system.pilot_light_btuh, pilot_light_btuh)
+    end
   end
 
-  def _test_default_stove_values(heating_system, fan_watts, heating_capacity)
+  def _test_default_stove_values(heating_system, fan_watts, heating_capacity, pilot_light, pilot_light_btuh)
     assert_equal(fan_watts, heating_system.fan_watts)
     if heating_capacity.nil?
       assert(heating_system.heating_capacity > 0)
     else
       assert_equal(heating_system.heating_capacity, heating_capacity)
+    end
+    assert_equal(heating_system.pilot_light, pilot_light)
+    if pilot_light_btuh.nil?
+      assert_nil(heating_system.pilot_light_btuh)
+    else
+      assert_equal(heating_system.pilot_light_btuh, pilot_light_btuh)
     end
   end
 
@@ -3680,12 +3768,18 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     end
   end
 
-  def _test_default_fireplace_values(heating_system, fan_watts, heating_capacity)
+  def _test_default_fireplace_values(heating_system, fan_watts, heating_capacity, pilot_light, pilot_light_btuh)
     assert_equal(fan_watts, heating_system.fan_watts)
     if heating_capacity.nil?
       assert(heating_system.heating_capacity > 0)
     else
       assert_equal(heating_system.heating_capacity, heating_capacity)
+    end
+    assert_equal(heating_system.pilot_light, pilot_light)
+    if pilot_light_btuh.nil?
+      assert_nil(heating_system.pilot_light_btuh)
+    else
+      assert_equal(heating_system.pilot_light_btuh, pilot_light_btuh)
     end
   end