+++ Changelog.md
@@ -11,6 +11,7 @@ __New Features__
   - Provides a warning if `BackupHeatingSwitchoverTemperature` or `BackupHeatingLockoutTemperature` are low and may cause unmet hours.
 - LightingGroups can now be specified using kWh/year annual consumption values as an alternative to fractions of different lighting types.
 - Allows building air leakage to be specified using CFMnatural or EffectiveLeakageArea.
+- Allows modeling a pilot light for non-electric heating systems (furnaces, stoves, and fireplaces).
 - Allows modeling one or more occupant vacancy periods (`VacancyPeriods` in the HPXML file).
 - Window shading seasons now based on calendar dates (e.g., summer: May 1-Sep 30 in northern hemisphere) instead of monthly outdoor temperatures.
 - ReportSimulationOutput measure: